# gameadapter/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/gamepool/CC/game/bet": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Probability Game Flow"
                ],
                "summary": "Probability Game Bet (遊戲回合下注)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "玩家遊戲token",
                        "name": "gametoken",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "遊戲代號",
                        "name": "gamecode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "遊戲回合編號",
                        "name": "round",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "下注金額※最大長度為12位數，及小數點後4位",
                        "name": "amount",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "交易代碼",
                        "name": "mtcode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "下注時間(UTC-4)",
                        "name": "datetime",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GameBetResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/bet/refund": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Table Flow"
                ],
                "summary": "Probability / Table Game Refund (下注取消)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "交易代碼",
                        "name": "mtcode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "玩家id",
                        "name": "userid",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TableGameRefundResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/currency": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Infos"
                ],
                "summary": "Currency List (幣別列表)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CurrencyListResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/detailtoken": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Order Detail(Game Result)"
                ],
                "summary": "Detail Token (驗證細單Token正確性並取回相關細單資訊)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "細單連結token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DetailTokenResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/end": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Probability Game Flow"
                ],
                "summary": "Probability Game End (遊戲回合結算)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "玩家遊戲token",
                        "name": "gametoken",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "遊戲代號",
                        "name": "gamecode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "遊戲回合編號",
                        "name": "round",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "jackpot金額",
                        "name": "jackpot",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "jackpot種類",
                        "name": "jackpottype",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "number"
                        },
                        "collectionFormat": "csv",
                        "description": "jackot貢獻值",
                        "name": "jackpotcontriubtion",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "傳送此Slot\tgame的FreeGame,LuckyDraw,Bonus的場次",
                        "name": "detail",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "結算時間(UTC-4)",
                        "name": "datetime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "number"
                        },
                        "collectionFormat": "csv",
                        "description": "殺率",
                        "name": "rtp",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "是否為再旋轉形成的注單",
                        "name": "singlerowbet",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "客端ip",
                        "name": "ip",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GameEndResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/generateoneroundcache": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Resend mechanism Flow"
                ],
                "summary": "Recover Gametoken by RoundID (當重送 End 時，遇到 error code 9 Invalid GameToken.，即可使用這個API重新激活相對應)",
                "parameters": [
                    {
                        "type": "string",
                        "default": "CC123456ab05:cq9",
                        "description": "訂單索引-\u003e roundid:xxx",
                        "name": "indexid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RecoverGametokenByRoundIDResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/generateroundcache": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Resend mechanism Flow"
                ],
                "summary": "Recover Gametoken by time range (當重送end時，遇到 error code 9 Invalid GameToken,即可使用這個 API 重新激活相對應時間區間內的 gametoken)",
                "parameters": [
                    {
                        "type": "string",
                        "default": "2023-07-01T00:00:00.000-04:00",
                        "description": "開始時間",
                        "name": "fromdate",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "2023-07-24T00:00:00.000-04:00",
                        "description": "結束時間(需小於現在時間一個小時)",
                        "name": "todate",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RecoverGametokenByTimeRangeResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/getorder": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Resend mechanism Flow"
                ],
                "summary": "Get Order (查詢注單內容)",
                "parameters": [
                    {
                        "type": "string",
                        "default": "CC123456ab07:cq9",
                        "description": "訂單索引-\u003e roundid:xxx",
                        "name": "indexid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetOrderResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/playerorder": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Others"
                ],
                "summary": "Player Order URL(取得玩家注單網址)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家遊戲token",
                        "name": "gametoken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PlayerOrderResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/promo": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Others"
                ],
                "summary": "Promotion(檢查目前是否有活動列表)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家遊戲token",
                        "name": "gametoken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PromotionResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/promo/link": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Others"
                ],
                "summary": "Promotion Link(產生活動連結)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "遊戲token",
                        "name": "gametoken",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "活動id",
                        "name": "promoid",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PromotionLinkResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/roundcheck": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "1.需於每個小時，定時使用 Round Check 對未完成的回合進行補單、退款行為。 ※每筆資料可以使用Round Detail、Get Order 得知我司狀態 舉例: 當 09:00 時 撈取 07:00~08:00，並每小時以此類推 對區間內的數據做補單、退款。\n※若貴司的紀錄為未成單卻在 Get order 撈取到該單據請直接使用注單內之資料將該注單成單\n2.補單流程是持續性的機制，須執行至所有未完成的回合成功，流程最少須持續24小時。 舉例: 當 09:00 時 執行 07:00~08:00 補單，並有回合未完成情況下，須至下次補單時繼續執行，直到該回合成功或是24小時後方可停止對該回合的補單。\n3.執行補單時，需和當下時間保持一定\"時間區間\"(依照遊戲自身演繹時間進行推算)，避免對當下進行中遊戲執行補單和退款。\n4.補單的datetime請使用補單當下時間",
                "tags": [
                    "Resend mechanism Flow"
                ],
                "summary": "Round Check (查詢未完成注單 (RoundID 必須符合格式才能查詢到))",
                "parameters": [
                    {
                        "type": "string",
                        "default": "2023-07-01T00:00:00.000-04:00",
                        "description": "開始時間",
                        "name": "fromdate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "2023-07-24T00:00:00.000-04:00",
                        "description": "結束時間",
                        "name": "todate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RoundCheckResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/game/rounddetail": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "依照查看到的狀態去做相對應的動作\n*Bet complete、init，且你的遊戲DB沒有 Win 的資料或相對應的後續 ➔ Refund\n*有Bet complete 且 Win init，且你的遊戲DB有相對應的資料 ➔ End\n*Rollout complete、init，但你的遊戲DB沒有相對應資料 ➔ refund\n*Rollout complete 且 Rollin init，且你的遊戲DB有相對應的資料 ➔ Rollin\n*Bet / Rollout failure 不執行任何動作\n*Bet / Rollout status 為 refunded 不執行任何動作",
                "tags": [
                    "Resend mechanism Flow"
                ],
                "summary": "Round Detail (查詢注單內容)",
                "parameters": [
                    {
                        "type": "string",
                        "default": "CC12334ab:cq9",
                        "description": "訂單索引-\u003e roundid:xxx",
                        "name": "indexid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RoundDetailResponse"
                        }
                    }
                }
            }
        },
...
        "/gamepool/CC/game/win/cancel": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Probability Game Flow"
                ],
                "summary": "Probability Game Win Cancel (贏分取消)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "交易代碼",
                        "name": "mtcode",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GameWinCancelResponse"
                        }
                    }
                }
            }
        },
...
        "/gamepool/CC/player/auth": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "ownerid = parentid 表示此parentid為總代理 ownerid ≠ parentid 表示此parentid為子代理 請務必紀錄parentid與ownerid，否則將產生對帳的疑慮",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Auth (驗證 Game Token，並且回傳此玩家的所有設定資訊)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "玩家遊戲token",
                        "name": "gametoken",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AuthPlayerResponse"
                        }
                    }
                }
            }
        },
        "/gamepool/CC/player/balance": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Balance (查詢餘額)",
                "parameters": [
                    {
                        "type": "string",
                        "default": "5995123703a236000175b842",
                        "description": "玩家id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "CC01",
                        "description": "遊戲代號",
                        "name": "gamecode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ShowBalanceResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.GameBetResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "資料給予的地方",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handler.GameBetResponseData"
                        }
                    ]
                },
                "status": {
                    "description": "狀態欄",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handler.GaHttpResponseStatus"
                        }
                    ]
                }
            }
        },
        "handler.GameBetResponseData": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "玩家餘額",
                    "type": "number"
                },
                "currency": {
                    "description": "玩家幣別",
                    "type": "string"
                }
            }
        },
...
        "handler.TableGameRefundResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "資料給予的地方",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handler.TableGameRefundResponseData"
                        }
                    ]
                },
                "status": {
                    "description": "狀態欄",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handler.GaHttpResponseStatus"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "輸入Bearer {jwtToken}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}